Upgrades.
* Scala 2.12.5 -> 2.12.6
* sbt-crossproject 0.3.0 -> 0.5.0 (for Scala.js 1.0.0-M5 support)
* Scala.js 0.6.23 -> 0.6.24
* Add Scala.js 1.0.0-M5

Support for Scala.js 1.0.0-M5 required to add some `.toInt`s to the
result of methods on `js.Date`, since they became `Double`s
upstream.

+++ src/main/scala/java/time/LocalDate.scala
@@ -391,7 +391,7 @@ object LocalDate {
 
   def now(): LocalDate = {
     val d = new js.Date()
-    of(d.getFullYear, d.getMonth + 1, d.getDate)
+    of(d.getFullYear.toInt, d.getMonth.toInt + 1, d.getDate.toInt)
   }
 
   // Not implemented
